Patch Powerline font patcher for Ricty glyph size
The Powerline font patcher normally scales the Powerline glyphs to match
the largest width and height of any glyph in the Latin-1 extended or
box-drawing ranges. This is far too big for Ricty, so this patch
overrides the font patcher with a custom glyph bounding box
(0, -525, 1025, 1650) that works well with Ricty.