Patch Discord script to support Inconsolata-dz
Inconsolata and Inconsolata-dz have alternate characters at the end of
the font that `ricty_discord_patch.pe` copies into the ASCII range by
default. However, the alternate characters are at different locations in
each font and the Ricty Discord patch assumes Inconsolata, so it copies
the wrong characters when Inconsolata-dz is used.

This change adds a patch (replacement) to `ricty_discord_patch.pe` to
work around the Ricty issue and manually specify the code points for the
alternate characters in Inconsolata-dz.

Fixes #17